Move around source files (#976)
* Move cli.js to a bin folder
* Move source files to src folder
* Move CLI code to its own file
