+++ src/infer.js
