+++ src/platform.js
