+++ src/prune.js
